Strategy_io can print its loaded strategy and also print auto now?
Feel free to reject if you think this is bad, but the goal of this commit is to make strategy_io called *after* --auto chooses a strategy. That way  --auto strategies can be printed. I also swapped the order of parsing and printing in strategy_io so that a parsed strategy can be printed to confirm it was processed.